+++ scaling/scale-up/topology/RingTopology.cpp
@@ -9,7 +9,7 @@ RingTopology::RingTopology(Bridge* bridge) : Topology(bridge) {}
 std::vector<Edge<std::string>> RingTopology::getBridgeEdges(std::vector<Graph*> samples) {
     std::vector<Edge<std::string>> bridges;
 
-    for (int i = 0; i < samples.size() - 1; i++) {
+    for (long long i = 0; i < samples.size() - 1; i++) {
         bridge->addBridgesBetweenGraphs(samples[i], samples[i + 1], bridges);
     }
